v0.2/requirements.txt
opencv-python
opencv-python-headless
onvif-zeep
pillow
numpy
appdirs